app/requirements.txt
fastapi==0.89.1
pydantic==1.10.4
python-dotenv==0.21.0
python-multipart==0.0.5
uvicorn==0.20.0
vaderSentiment==3.3.2
loguru==0.6.0